requirements.txt
pymupdf
spacy==3.8.11
requests
jinja2
pyyaml
